+++ app/Http/Controllers/Gescon/ContratoCrudController.php
@@ -261,6 +261,17 @@ class ContratoCrudController extends CrudController
                 'visibleInShow' => true, // sure, why not
             ],
             [
+                'name' => 'formatVlrAcumulado',
+                'label' => 'Valor Acumulado', // Table column heading
+                'type' => 'model_function',
+                'function_name' => 'formatVlrAcumulado', // the method in your Model
+                'orderable' => true,
+                'visibleInTable' => false, // no point, since it's a large text
+                'visibleInModal' => true, // would make the modal too big
+                'visibleInExport' => true, // not important enough
+                'visibleInShow' => true, // sure, why not
+            ],
+            [
                 'name' => 'situacao',
                 'label' => 'Situação',
                 'type' => 'boolean',